Therapify/requirements.txt
chainlit
openai
python-dotenv
requests